% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_color_opt.R
\name{act_color_opt}
\alias{act_color_opt}
\title{act_color_opt}
\usage{
act_color_opt()
}
\value{
names of available pre-fab colors
}
\description{
act_color_opt
}
\examples{
act_show_colors("aladdin_sane")



}
