% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_extract_svydesign.R
\name{act_extract_svydesign}
\alias{act_extract_svydesign}
\title{Extracts dataframe from svydesign. Not terribly useful on own, but called in other Rtesunate functions}
\usage{
act_extract_svydesign(design_object)
}
\arguments{
\item{design_object}{svydesign object of interest in functions act_row, act_ns, etc}
}
\value{
Dataframe used to build svydesign used
}
\description{
Extracts dataframe from svydesign. Not terribly useful on own, but called in other Rtesunate functions
}
\examples{
using api data from survey package
act_extract_svydesign(apistrat)
}
