% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_join.R
\name{act_join}
\alias{act_join}
\title{Modifier of tidyverse joins to return only unique 1-to-1 joins or a list of
dataframes serving for join 'diagnostics'}
\usage{
act_join(x, y, type = "left_join", out = "pipe", ...)
}
\arguments{
\item{x}{a dataframe to join}

\item{y}{another dataframe to join}

\item{type}{type of join to perform (in quotes). Options are "left_join" (default),
"inner_join", "full_join", "semi_join", and "right_join". See separate help
pages for these functions to determine which is most appropriate}

\item{out}{output desired (in quotes). "pipe" (default) returns a single dataframe
containing only perfect 1-to-1 matches. "list" returns a list of dataframes
containing all 'diagnostic' categories (1-to-1, rows in x duplicated, rows
in y duplicated, unmatched observations from x, unmatched observations from y)}

\item{...}{passed to join function of choice}
}
\value{
Dependent on 'out' parameter. See above
}
\description{
Modifier of tidyverse joins to return only unique 1-to-1 joins or a list of
dataframes serving for join 'diagnostics'
}
\examples{
test <- act_join(x, y, by = c("param1" = "param1", "param2" = "param2"))

}
