% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_ns.R
\name{act_ns}
\alias{act_ns}
\title{A creates a dataframe with unweighted n/N observation counts for value of interest
of selected variable within a svydesign (categorical). Best used in conjunction with Rtesunate::act_rows().
(See act_svyciprop family of functions which combines the two)}
\usage{
act_ns(x, cond, design)
}
\arguments{
\item{x}{variable of interest (in quotes)}

\item{cond}{the value for which estimates are desired (in quotes)}

\item{design}{svydesign object containing the variable/value}
}
\value{
A dataframe providing the counts as n/N character variable (N is all non-NA values)
}
\description{
A creates a dataframe with unweighted n/N observation counts for value of interest
of selected variable within a svydesign (categorical). Best used in conjunction with Rtesunate::act_rows().
(See act_svyciprop family of functions which combines the two)
}
\examples{
using api data from survey package
data(api)
design <- svydesign(id=~1,strata=~stype, weights=~pw, data=apistrat, fpc=~fpc)
act_ns("cname", "Los Angeles", design)
}
