% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_under_vect.R
\name{act_one_under}
\alias{act_one_under}
\title{act_one_under}
\usage{
act_one_under(x)
}
\arguments{
\item{x}{vector to assess for structure}
}
\value{
Converts character vectors containing entries that all have one or
fewer underscores to numeric after substituting a decimal for the underscore
}
\description{
act_one_under
}
\examples{
num_vector <- act_one_under(x)
}
