% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_pal.R
\name{act_pal}
\alias{act_pal}
\title{Generates a function to create a palette based on those included in the
package.}
\usage{
act_pal(palette = "ncezid", reverse = FALSE)
}
\arguments{
\item{palette}{name of pre-fab palette on which to base output (in quotes)}

\item{reverse}{boolean variable defaulting to FALSE. Set to TRUE if want to
reverse}

\item{...}{arguments passed to colorRampPalette}
}
\value{
A function which outputs a palette of desired length, interpolating
between stated colors if necessary.
}
\description{
Generates a function to create a palette based on those included in the
package.
}
\examples{
pal_5 <- act_pal("aladdin_sane")(5)
}
