% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_row.R
\name{act_row}
\alias{act_row}
\title{A creates a dataframe with weighted means and cluster adjusted 95\%
confidence intervals for specific responses for categorical data in svydesign}
\usage{
act_row(x, term, design)
}
\arguments{
\item{x}{variable of interest (in quotes)}

\item{term}{the value for which estimates are desired (in quotes)}

\item{design}{svydesign object containing the variable/value}
}
\value{
A dataframe providing the mean, 95\% CI
}
\description{
A creates a dataframe with weighted means and cluster adjusted 95\%
confidence intervals for specific responses for categorical data in svydesign
}
\examples{
using api data from survey package
data(api)
design <- svydesign(id=~1,strata=~stype, weights=~pw, data=apistrat, fpc=~fpc)
act_row("cname", "Los Angeles", design)
}
