% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_row_quant.R
\name{act_row_quant}
\alias{act_row_quant}
\title{Calculates weighted mean and cluster-adjusted confidence interval for quantitative
variable in svydesign object}
\usage{
act_row_quant(x, design, ...)
}
\arguments{
\item{x}{variable of interest (in quotes)}

\item{...}{arguments passed to stats::confint. Most useful if need to change
from 95\% CI (default) to a different level of confidence; entered as
'level = 0.xx'}

\item{design_object}{svydesign object of interest in functions act_row, act_ns, etc}
}
\value{
Dataframe with one row indicating variable name, survey-weighted mean
and confidence interval
}
\description{
Calculates weighted mean and cluster-adjusted confidence interval for quantitative
variable in svydesign object
}
\examples{
using api data from survey package
design <- svydesign(id=~1,strata=~stype, weights=~pw, data=apistrat, fpc=~fpc)
act_row_quant("api00", design)
act_row_quant("api00", design, level = 0.99)

}
