% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_scale_color.R
\name{act_scale_color}
\alias{act_scale_color}
\title{Colors ggplots with palettes based on pre-fab palettes included in this
package, at no extra cost!}
\usage{
act_scale_color(palette = "ncezid", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{name of pre-fab palette on which to base output (in quotes)}

\item{reverse}{boolean variable defaulting to FALSE. Set to TRUE if want to
reverse}

\item{...}{arguments passed to discrete_scale or scale_fill_gradientn, which
do the work under the hood.}
}
\value{
A ggplot with colors that ROCK (or that are sort of boring, depending
on the pre-fab palette chosen).
}
\description{
Colors ggplots with palettes based on pre-fab palettes included in this
package, at no extra cost!
}
\examples{
data("iris")

library(tidyverse)

ggplot(iris)+
 geom_point(aes(x = Sepal.Width, y = Petal.Length, color = Species))+
 act_scale_color("hello_nasty")
}
