% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_show_colors.R
\name{act_show_colors}
\alias{act_show_colors}
\title{act_show_colors}
\usage{
act_show_colors(palette)
}
\arguments{
\item{palette}{Palette name (in quotes)}
}
\value{
Righteous colors
}
\description{
Let's ROCK!
}
\examples{
get_jams()

}
