% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_svyciprop.R
\name{act_svyciprop}
\alias{act_svyciprop}
\title{A creates a one line dataframe with weighted mean, cluster-adjusted 95\% CI, and
unweighted n/N observation counts for value of interest for selected variable
within a svydesign.}
\usage{
act_svyciprop(x, cond, design, ...)
}
\arguments{
\item{x}{variable of interest (in quotes)}

\item{cond}{the value for which estimates are desired (in quotes)}

\item{design}{svydesign object containing the variable/value}

\item{...}{arguments passed to svyciprop}
}
\value{
A dataframe providing the variable name, weighted mean proportion of
observations taking value of interest, cluster-adjusted 95\% CI, and counts
as n/N character variable (N is all non-NA values).
}
\description{
A creates a one line dataframe with weighted mean, cluster-adjusted 95\% CI, and
unweighted n/N observation counts for value of interest for selected variable
within a svydesign.
}
\examples{
using api data from survey package
data(api)
design <- svydesign(id=~1,strata=~stype, weights=~pw, data=apistrat, fpc=~fpc)
act_svyciprop("cname", "Los Angeles", design)
}
