% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_svyciprop_list.R
\name{act_svyciprop_list}
\alias{act_svyciprop_list}
\title{Listified version of act_svyciprop_tbl, allowing parallel use with multiple
svydesigns. Useful when survey conducted at multiple times and/or there are
multiple subsets of interest. Variables and responses are input as vectors;
svydesigns are input as lists}
\usage{
act_svyciprop_list(variable_vect_name, response, design_obj_list)
}
\arguments{
\item{variable_vect_name}{a vector of variables of interest (in quotes)}

\item{response}{a vector of response of intersest for each entry in varibale_vect_name.
Note that these are paired by order and that the number of entries in
variable_vect_name and response must be equal}

\item{design_obj_list}{list of svydesign object containing the variables/values.}
}
\value{
A dataframe providing the variable names, weighted mean proportions of
observations taking value of interest, cluster-adjusted 95\% CIs, and counts
as n/N character variable (N is all non-NA values). Results from different
svydesigns appear as columns in the order in which they are called.
}
\description{
Listified version of act_svyciprop_tbl, allowing parallel use with multiple
svydesigns. Useful when survey conducted at multiple times and/or there are
multiple subsets of interest. Variables and responses are input as vectors;
svydesigns are input as lists
}
\examples{
using api data from survey package
data(api)
design <- svydesign(id=~1,strata=~stype, weights=~pw, data=apistrat, fpc=~fpc)
design1 <- subset(design, cname == "Los Angeles")
design2 <- subset(design, cname == "San Diego")
variable_vect_name <- c("comp.imp", "sch.wide", "awards")
response <- c("Yes", "Yes", "Yes")
act_svyciprop_list(variable_vect_name, response, list(design1, design2))
}
