% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_svyciprop_tbl.R
\name{act_svyciprop_tbl}
\alias{act_svyciprop_tbl}
\title{Similar to act_svyciprop, but able to spit out a dataframe containing estimates
for multiple variables/responses. Variables and responses are input as vectors.}
\usage{
act_svyciprop_tbl(variable_vect_name, response, design_obj, ...)
}
\arguments{
\item{variable_vect_name}{a vector of variables of interest (in quotes)}

\item{response}{a vector of response of intersest for each entry in varibale_vect_name.
Note that these are paired by order and that the number of entries in
variable_vect_name and response must be equal}

\item{...}{arguments passed to svyciprop}

\item{design}{svydesign object containing the variables/values. Can only take one svydesign.
To call multiple svydesigns in parallel, see purrr package or Rtesunate::act_svyciprop_list()}
}
\value{
A mutli-row dataframe providing the variable names, weighted mean proportions of
observations taking value of interest, cluster-adjusted 95\% CIs, and counts
as n/N character variable (N is all non-NA values).
}
\description{
Similar to act_svyciprop, but able to spit out a dataframe containing estimates
for multiple variables/responses. Variables and responses are input as vectors.
}
\examples{
using api data from survey package
data(api)
design <- svydesign(id=~1,strata=~stype, weights=~pw, data=apistrat, fpc=~fpc)
variable_vect_name <- c("cname", "sch.wide", "awards")
response <- c("Los Angeles", "Yes", "Yes")
act_svyciprop_tbl(variable_vect_name, response, design)

}
