% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_svyciquant_list.R
\name{act_svyciquant_list}
\alias{act_svyciquant_list}
\title{Listified version of act_svyciquant_tbl, allowing parallel use with multiple
svydesigns. Useful when survey conducted at multiple times and/or there are
multiple subsets of interest. Variables are input as vectors;
svydesigns are input as lists}
\usage{
act_svyciquant_list(variable_vect_name, design_obj_list, ...)
}
\arguments{
\item{variable_vect_name}{a vector of variables of interest (in quotes)}

\item{design_obj_list}{list of svydesign object containing the variables/values.}

\item{...}{arguments passed to svyciprop}
}
\value{
A dataframe providing the variable names, weighted mean proportions of
observations taking value of interest, cluster-adjusted 95\% CIs, and counts
as n/N character variable (N is all non-NA values). Results from different
svydesigns appear as columns in the order in which they are called.
}
\description{
Listified version of act_svyciquant_tbl, allowing parallel use with multiple
svydesigns. Useful when survey conducted at multiple times and/or there are
multiple subsets of interest. Variables are input as vectors;
svydesigns are input as lists
}
\examples{
using api data from survey package
data(api)
design <- svydesign(id=~1,strata=~stype, weights=~pw, data=apistrat, fpc=~fpc)
design1 <- subset(design, cname == "Los Angeles")
design2 <- subset(design, cname == "San Diego")
variable_vect_name <- c("api00", "snum")
act_svyciquant_list(variable_vect_name, list(design1, design2))
act_svyciquant_list(variable_vect_name, list(design1, design2), level = 0.99)
}
