% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_xtab.R
\name{act_xtab}
\alias{act_xtab}
\title{Creates crosstabs showing weighted proportion and CIs for two categorical
variables of interest}
\usage{
act_xtab(x, y, design, hundo = "row", print = FALSE, ...)
}
\arguments{
\item{x}{variable of interest in svydesign object}

\item{y}{variable of interest in svydesign object}

\item{design}{svydesign object}

\item{hundo}{string variable indicating how estimates add to 100\%. Current
options are "diagonal" or "row" (the default)}

\item{print}{Boolean variable; determines if output is tidy/friendly for
additional manipulation or in 2x2 format for print/human reading. Defaults to
FALSE}
}
\value{
Crosstab of weighted proportions/CIs for two categorical variables
}
\description{
Creates crosstabs showing weighted proportion and CIs for two categorical
variables of interest
}
\examples{
library(survey)
data(api)
dclus1 <- svydesign(id=~dnum, weights=~pw, data=apiclus1, fpc=~fpc)

act_xtab("sch.wide", "comp.imp", dclus1, print = TRUE, method = "li")
}
