% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jams.R
\name{get_jams}
\alias{get_jams}
\title{get_jams}
\usage{
get_jams(secret_code = NULL)
}
\arguments{
\item{secret_code}{Accepts a secret code, if you know any (in quotes)}
}
\value{
Righteous tunes
}
\description{
Let's ROCK!
}
\examples{
get_jams()
}
