% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squeaky_clean.R
\name{squeaky_clean}
\alias{squeaky_clean}
\title{squeaky_clean}
\usage{
squeaky_clean(db, indiv_num = NULL)
}
\arguments{
\item{db}{Database to clean}

\item{indiv_num}{Vector of variables in format of c("x", "y", etc) to retain as numeric if all whole numbers}
}
\value{
A hopefully cleaner database
}
\description{
squeaky_clean
}
