% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tes_data.R
\docType{data}
\name{tes_data}
\alias{tes_data}
\title{TES data}
\format{
A data frame with 346 rows and 39 columns:
\describe{
\item{Country}{Country name}
\item{Site}{Study site name}
\item{Study Pop}{Population in which the study conducted}
\item{Start Year, End Year, Year of Study Report}{When study conducted/reported}
\item{Arm (Antimalarial)}{Medication administered in the study}
\item{First Line Anitmalarial}{First line malaria medication per national guidelines at the time of the study}
\item{PerProtocol_D28_Uncorrected}{Efficacy at day 28, uncorrected}
\item{PerProtocol_CI_D28_Uncorrected_low}{Lower value of 95\% confidence interval for D28 uncorrected}
\item{PerProtocol_CI_D28_Uncorrected_high}{Higher value of 95\% confidence interval for D28 uncorrected}
\item{PerProtocol_D28_Corrected}{Efficacy at day 28, PCR corrected}
\item{PerProtocol_D28_CI_Corrected_low}{Lower value of 95\% confidence interval for D28 corrected}
\item{PerProtocol_D28_CI_Corrected_high}{Higher value of 95\% confidence interval for D28 corrected}
\item{PerProtocol_D42_Uncorrected}{}
\item{PerProtocol_CI_D42_Uncorrected_low}{}
\item{PerProtocol_CI_D42_Uncorrected_high}{}
\item{PerProtocol_D42_Corrected}{}
\item{PerProtocol_CI_D42_Corrected_low}{}
\item{PerProtocol_CI_D42_Corrected_high}{}
\item{Kaplan-Meier day 28_Uncorrected}{}
\item{Kaplan_Meier_d28_CI_Uncorrected_low}{}
\item{Kaplan_Meier_d28_CI_Uncorrected_high}{}
\item{Kaplan-Meier Day 28_corrected}{}
\item{Kaplan_Meier_d28_CI_corrected_low}{}
\item{Kaplan_Meier_d28_CI_corrected_high}{}
\item{Kaplan-Meier day 42_Uncorrected}{}
\item{Kaplan-Meier Day 42_corrected}{}
\item{Genotype markers}{}
\item{Raw Genotyping Available}{}
\item{Reinfections included in numerator of corrected efficacy}{}
\item{Day of efficacy results}{}
\item{PercentACPR_Corrected}{}
\item{ConfidenceInt_Corrected}{}
\item{MOP Funding Year}{}
\item{Funder}{}
\item{report_link}{}
\item{Lat}{}
\item{Long}{}
}
\usage{
tes_data
}
\description{
Data from TES performed in  PMI partner countries, as well as select
studies non-PMI countries. Updated 16 May, 2023
}
\examples{
data <- tes_data

}
\keyword{datasets}
