% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_fullmonty.R
\name{act_fullmonty}
\alias{act_fullmonty}
\title{A dataframe with weighted mean, cluster-adjusted 95\% CI, and
unweighted n/N observation counts for every variable
within a svydesign.}
\usage{
act_fullmonty(design, drop = NULL)
}
\arguments{
\item{design}{svydesign object containing the variable/value}

\item{drop}{vector of variables to drop from output dataframe}
}
\value{
A dataframe providing the variable name, value, weighted mean proportion of
all observations, cluster-adjusted 95\% CI, and counts
as n/N character variable (N is all non-NA values).
}
\description{
A dataframe with weighted mean, cluster-adjusted 95\% CI, and
unweighted n/N observation counts for every variable
within a svydesign.
}
\examples{
using api data from survey package
data(api)
design <- svydesign(id=~1,strata=~stype, weights=~pw, data=apistrat, fpc=~fpc)
act_svyciprop("cname", "Los Angeles", design)
}
